Bump sway to v1.7
This updates wayland to 1.20, allowing us to compile sway 1.7. Unfortunately, we're not exactly on wlroots 0.15.0: a couple of commits before 0.15 they bumped libdrm requirement to 2.4.109 (we have 2.4.107). This lib is too dangerous to backport from jammy.

Because of libdrm's requirement, this is probably as far as we can get with Ubuntu 21.10. Any further updates will require Ubuntu 22.04.

Changes:

 * Add libwayland 1.20 libs from ubuntu jammy
 * Bump wlroots version to "almost" 0.15.0
 * Bump sway version to 1.27